pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lana-scenario-gen"
version = "0.1.0"
description = "Generate SQL event data for lana-bank from scenario definitions"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "GaloyMoney" }]
keywords = ["lana", "events", "sql", "testing", "scenarios"]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "psycopg[binary]>=3.1",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[project.scripts]
lana-scenario = "lana_scenario_gen.cli:main"

[project.urls]
Repository = "https://github.com/GaloyMoney/lana-scenario-gen"

[tool.hatch.build.targets.wheel]
packages = ["src/lana_scenario_gen"]
